third_party/blink/web_tests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-pointsAtX-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtX' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.pointsAtX.baseVal is -30
PASS spotLightElement.pointsAtX.baseVal is 83
PASS successfullyParsed is true

TEST COMPLETE
